added dtype=int after the np.zeros two dimentional code so now all of the integer values are in fact integer instead of decimal points and values.

+++ Pascal_Triangle/result.py
@@ -1,7 +1,7 @@
 import numpy as np 
 
 def generate(number): 
-    newlist = np.zeros((number, number))
+    newlist = np.zeros((number, number), dtype=int)
     
     for i in range(number): 
         for j in range(i):
